requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
langchain-groq
langsmith
pypdf
faiss-cpu
transformers 
accelerate 
torch 
sentence-transformers

